% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{swellCalibration}
\alias{swellCalibration}
\title{Run SWELL Calibration}
\usage{
swellCalibration(
  weather_data,
  vegetation_data,
  vegetationIndex = "EVI",
  SWELLparameters,
  start_year = 2011,
  end_year = 2022,
  simplexes = 1,
  iterations = 1
)
}
\arguments{
\item{weather_data}{A data frame containing geographical and weather data.
Columns must be: Latitude (numeric), Longitude (numeric), Date (YYYY-MM-DD), Tmin (numeric, °C), Tmax (numeric, °C).}

\item{vegetation_data}{A data frame containing the vegetation index data.
Columns must be: PixelID (string), Group (string), Year (integer), Doy (integer), Longitude (numeric), Latitude (numeric), VegetationIndex (numeric).}

\item{SWELLparameters}{A data frame containing SWELL model parameters.
Columns must be: species (string), class (string), parameter (string), min (float), max (float), value (float), is_calibrated (boolean). See 'SWELLparameters' object for the template.}

\item{start_year}{Start year for calibration (default: 2011).}

\item{end_year}{End year for calibration (default: 2022).}

\item{simplexes}{Number of simplexes for calibration (default: 3).}

\item{iterations}{Number of iterations for calibration (default: 1000).}

\item{vegetation_index}{The vegetation index used for SWELL calibration (string). Available options are 'EVI' and 'NDVI'}
}
\value{
A list containing:
\itemize{
\item calibration_results: A data frame combining all calibration results with SWELL outputs.
\item parameters_pixels: A data frame containing the calibrated values of SWELL parameters.
\item parameters_group: A data frame containing the mean and standard deviation of SWELL parameters grouped by the Group column in the vegetation_data dataframe
}
}
\description{
This function performs the calibration of the SWELL model using weather data as input and vegetation indices data (NDVI or EVI) as reference data using a multi-start simplex algorithm.
}
