% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWELLparameters.R
\docType{data}
\name{SWELLparameters}
\alias{SWELLparameters}
\title{SWELL parameters data used in calibration}
\format{
A data frame with 31 rows and 7 columns:
\describe{
\item{species}{Species name.}
\item{class}{Class of the parameter.}
\item{parameter}{Name of the parameter.}
\item{min}{Minimum value of the parameter (numeric). Used for automatic calibration if \code{calibration} is TRUE.}
\item{max}{Maximum value of the parameter (numeric). Used for automatic calibration if \code{calibration} is TRUE.}
\item{value}{Default value of the parameter (numeric).}
\item{calibration}{Logical, TRUE if the parameter is under calibration, FALSE otherwise.}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
SWELLparameters
}
\description{
A dataframe providing default values for SWELL application to reproduce beech vegetation dynamics.
Users can modify this dataset as needed before using it in the \code{swellCalibration} function for calibration.
}
\details{
This dataset is used as the starting point for parameter values in the \code{swellCalibration} function.
Users can edit or replace parameter values to suit specific modeling needs or calibration goals.
}
\keyword{datasets}
